requirements.txt
setuptools~=57.0.0
python-magic~=0.4.24